Add time scaling and make fixed timestep better

+++ examples/2d/text2d.rs
@@ -94,8 +94,8 @@ fn animate_translation(
     mut query: Query<&mut Transform, (With<Text>, With<AnimateTranslation>)>,
 ) {
     for mut transform in query.iter_mut() {
-        transform.translation.x = 100.0 * time.seconds_since_startup().sin() as f32 - 400.0;
-        transform.translation.y = 100.0 * time.seconds_since_startup().cos() as f32;
+        transform.translation.x = 100.0 * time.seconds_since_startup().sin() - 400.0;
+        transform.translation.y = 100.0 * time.seconds_since_startup().cos();
     }
 }
 
@@ -104,7 +104,7 @@ fn animate_rotation(
     mut query: Query<&mut Transform, (With<Text>, With<AnimateRotation>)>,
 ) {
     for mut transform in query.iter_mut() {
-        transform.rotation = Quat::from_rotation_z(time.seconds_since_startup().cos() as f32);
+        transform.rotation = Quat::from_rotation_z(time.seconds_since_startup().cos());
     }
 }
 
@@ -116,6 +116,6 @@ fn animate_scale(
     // rendered quad, resulting in a pixellated look.
     for mut transform in query.iter_mut() {
         transform.translation = Vec3::new(400.0, 0.0, 0.0);
-        transform.scale = Vec3::splat((time.seconds_since_startup().sin() as f32 + 1.1) * 2.0);
+        transform.scale = Vec3::splat((time.seconds_since_startup().sin() + 1.1) * 2.0);
     }
 }
